pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stonefish"
version = "0.1.0"
description = "Stonefish project"
readme = "readme.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "torch>=2.0.0",
    "transformers[flax]",
    "datasets",
    "python-chess",
    "pytest",
    "pytest-cov",
    "rich",
    "matplotlib",
    "pyyaml",
    "chessplotlib",
    "yamlargs",
    "mllg",
    "optax",
    "wandb",
]

[project.scripts]

[tool.setuptools]
packages = ["stonefish", "stonefish.eval", "stonefish.train"]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "ruff",
]
test = [
    "pytest",
    "pytest-cov",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "N", "W", "B"]
ignore = []

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = "test_*.py"
